+++ src/lightning_app/cli/cmd_init.py
@@ -124,7 +124,7 @@ def component(component_name):
     ⚡ Use the component inside an app: ⚡
 
     from {name_for_files} import TemplateComponent
-    import lightning_app as la
+    import lightning.app as la
 
     class LitApp(la.LightningFlow):
         def __init__(self) -> None: